Move app component into components folder
